+++ src/views/login/index.vue
@@ -135,7 +135,7 @@ import loginMixin from './loginMixin.js'
 import SocialSign from './components/SocialSignin'
 
 // API Request Methods
-import { services } from '@/api/ADempiere/Security/index.ts'
+import { services } from '@/api/ADempiere/security//index.ts'
 
 // Utils and Helper Methods
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
